# Optimize client conversions for client-server array transfers