--- a/arkouda/dtypes.py
+++ b/arkouda/dtypes.py
@@ -3,12 +3,13 @@ from enum import Enum
 import numpy as np # type: ignore
 from typeguard import typechecked
 import builtins
+import sys
 
 __all__ = ["DTypes", "DTypeObjects", "dtype", "bool", "int64", "float64", 
            "uint8", "str_", "check_np_dtype", "translate_np_dtype", 
            "resolve_scalar_dtype", "ARKOUDA_SUPPORTED_DTYPES", "bool_scalars",
            "float_scalars", "int_scalars", "numeric_scalars", "numpy_scalars", 
-           "str_scalars", "all_scalars"]
+           "str_scalars", "all_scalars", "get_byteorder"]
 
 # supported dtypes
 structDtypeCodes = {'int64': 'q',
@@ -161,3 +162,17 @@ def resolve_scalar_dtype(val : object) -> str: # type: ignore
     # Other python type
     else:
         return builtins.str(type(val))
+
+def get_byteorder(dt: np.dtype) -> str:
+    """
+    Get a concrete byteorder (turns '=' into '<' or '>')
+    """
+    if dt.byteorder == '=':
+        if sys.byteorder == 'little':
+            return '<'
+        elif sys.byteorder == 'big':
+            return '>'
+        else:
+            raise ValueError("Client byteorder must be 'little' or 'big'")
+    else:
+        return dt.byteorder