Improve the performance of `ak.array()` and `pdarray.to_ndarray()` by
optimizing how the client converts between bytes and numpy arrays. This
switches from `struct.{pack,unpack}` to `numpy.{tobytes,frombuffer}`
which has a non-trivial performance improvement. We have to be more
careful about endianness now, but the conversions are still pretty
straightforward. Note that the server currently always uses big-endian
for these conversions though a future optimization will look to use
native endianness.

Currently, the client code is still making some extra copies. In the
`to_ndarray()` case we need a copy because the underlying `bytes` buffer
is immutable, which makes the numpy array immutable so things that
modify the array in place like `numpy.array.sort()` would fail without
the copy. We should be able to optimize this by looking at the zmq frame
directly, but this will be future work. On the `ak.array()` side I think
`tobytes` copies, and I wonder if we can use `memoryview` here instead.
That too will be left for future work.

Here's the performance improvement for 16-node-xc:

| config | to_ndarray | ak.array   |
| ------ | ---------: | ---------: |
| before | 20.5 MiB/s | 23.5 MiB/s |
| after  | 32.5 MiB/s | 50.0 MiB/s |

While the rate increases don't look too large the time difference shows
the improvement a little better.

| config | to_ndarray | ak.array   |
| ------ | ---------: | ---------: |
| before | 36.6s      | 31.2s      |
| after  | 23.6s      | 15.2s      |

Next steps will be to optimize the server side conversions more, which
should shave another 10-15 seconds off these operations which will give
more respectable rates.

Part of #794